Work on mutation matrices
